Abstract the period name builder
